Add Gemfile.lock
This file SHOULD NOT be ignored, because it is the only way to
ensure the bundle consitency accross different repositories.

It should never be edited by hand either, and any change to the
Gemfile or gemspec should be committed with the corresponding
Gemfile.lock - run 'bundle install' to update the lock.