Minor refactor remove assertion on class
Asserting on the objects class is not used much in Ruby, and
in this case it was used in case 'hooks' were nil.
Instead, 'hooks' could default to an empty array (which conveys
more meaning that nil) and it would always responds to "#each".

The solution only builds on top of the existing duplication,
but that's a different problem.