Refactor extract method from CLI to Server#disable_buffering
Ensuring that the output doesn't get buffered is particularly
relevant once the server is started because Dredd relies on it.
In order to be able to start programatically a server in
the best conditions, that step should be part of the server
starting procedure.

Using a buffered output while loading the hookfiles, on the other
hand poses no particular problem.