Refactor remove @debug tags from features
Aruba v0.7.0 did add support for a @debug tag that didn't exist before.
That @debug tag seems to allow for the use of debugging tools
to debug failures that only happen in the Cucumber features.
That also means that our features fail when using Aruba >= v0.7.0.

See https://github.com/cucumber/aruba/commit/959de81dd03f6ab12e770fb920ab718e0a53ec96

It also means that the @debug tag that was in place in our
features was not related to Aruba. Cucumber doesn't seem to assign
any specific meaning to the @debug tag either.

I couldn't track the origin of that tag before
https://github.com/apiaryio/dredd-hooks-template/commit/7573e351d2cf5033cba97f0bb25734e26425eb1f,
and I believe it might have been left in place after being used locally
to write the features in the first place. If that is true, removing
it would be the right thing to do in order to be able to update Aruba.

Note: when using Aruba v0.6.2, the @debug tag has in fact no visible
effect, so this commit is actually a refactoring!

See also https://github.com/apiaryio/dredd-hooks-ruby/compare/v0.1.2...gonzalo-bulnes:spike-update-aruba?expand=1